changed hw and LinkedList, DynamicArray

+++ sem-i-prakt/data-structures/generics/LinkedList/main.cpp
@@ -6,11 +6,11 @@ int main()
 {
     LinkedList<int> l;
 
-    for(int i=0;i<10;i++)
+    for (size_t i = 0; i < 10; i++)
         l.push_back(i);
 
-    for(int i=0;i<10;i++)
-        cout<<l[i]<<endl;
+    for (size_t i = 0; i < 10; i++)
+        cout << l[i] << endl;
 
 	return 0;
 }
